third_party/WebKit/LayoutTests/fast/dom/Document/replaceChild-null-oldChild.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
description('Test behavior of Document.replaceChild() when oldChild is null.');

shouldThrow('document.replaceChild(document.firstChild, null)', '"TypeError: Failed to execute \'replaceChild\' on \'Node\': parameter 2 is not of type \'Node\'."');
</script>
</body>
</html>
